requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
Pillow==10.4.0
aiofiles==24.1.0
